Upload V2.1 - StarID
StarID got a pretty major overhaul, restructuring into a logical order, and explaining everything. It is also now up to spec with path integration, and has the parameters fairly set-able. The QAOP_starID module contains only the autonamer class, and not the full process, so that will remain for a V3 maybe.