Upload V2.2
This version finishes updating the path integration throughout the module (I think). RadialProfiling hasn't been fully overhauled, but some of the initial changes were made. It does need to be looked at again after the StarID overhaul though, and have a similar thing done to it. Also on the back burner is making a new-StarID-esque demo notebook for the photometry class.